Mark components with a menu item number of 99999 or 99900 as uninstallable
This can be used for components that should be selectable by users in anna,
but don't have functionality that can be run from the main menu. The change
avoids the need to add a dummy isinstallble script for such components.
Examples are openssh-client and fdisk.

The use of 99900 is deprecated, but needed as components currently use that
number. It can be removed once they have been changed to use 99999.