Test sources before adding them to sources.list.udeb
We currently assume that all sources in the system's /etc/apt/sources.list
are valid for the target suite, which may not be true if it contains an
incomplete mirror. Example is the lxdebian s390 build which lists a
mirror that does not have testing and thus fails.

Solve this by testing that a mirror actually contains the main/d-i section
for the target suite.