fix Symbol.map moving code; revert workaround for 2nd-order breakage
This was broken by r60622, adding FreeBSD support, which chose the wrong
variable to determine if building for linux.

The 1st order breakage from that is the System.map file was not being used
when running depmod to generate modules.dep, which can be a problem if
the running kernel's System.map doesn't match the system being built.

The 2nd order breakage was that it left System.map in boot/ and broke the
builds.