get-packages: correctly restore original Packages lists when script exits
Turns out the script was buggy due to the use of a confusing variable name.
The bug could result in Packages lists not being restored correctly when
the script exits if multiple Packages lists are present.

This probably explains the build failures on armel for 20091215. It's not
clear why only armel would be affected.

Apparently APT itself does not notice when these files are "corrupted".