Add check for incorrect file attributes on device nodes
This can occur when using fakeroot after restarting a (failed) build.
See fakeroot(1) ('-s' and '-i' options) for background.

As this has caused me hours of wasted time while trying to port D-I
to an ARM netbook, it is better to fail the build than produce broken
images even if the missing file attributes will be harmless in practice
in most cases.