armhf: Select armp kernel when appropriate.
For $SUBARCH == "generic" simply return "armmp".

For $SUBARCH other than "generic" return both the subarch (which will
correspond to a kernel flavour in Wheezy and earlier) and armmp. This means
that the installer should still be usable for installing Wheezy (and perhaps
earlier).

Better support multiple kernel flavours (code cribbed from i386). Extend test
suite with an example of a "generic" machine.